fix: hex: 'Enter in Hex' mode: 0x80 이상의 문자열을 취급하지 못하는 문제 수정

+++ src/main/java/org/gosqo/tcpsocket/HexConverter.java
@@ -42,7 +42,7 @@ public class HexConverter {
                 throw new IllegalArgumentException("cannot pair in 2 digit hex");
             }
 
-            bytes[i] = Byte.parseByte(each, 16); // NumberFormatException
+            bytes[i] = (byte) Integer.parseInt(each, 16); // NumberFormatException
         }
 
         return bytes;
